Add files via upload
This analysis breaks the misconception that location of a business effects the rating.
The location considered for the analysis is Las Vegas. So, had that been the case, all the businesses near The Strip would have higher ratings as compared to other locations in the city. Which is not the case as discovered in the analysis.